Move StaleBlockchain to exceptions module
