+++ web3/exceptions.py
@@ -1,3 +1,7 @@
+import datetime
+import time
+
+
 class BadFunctionCallOutput(Exception):
     """
     We failed to decode ABI output.
@@ -20,3 +24,17 @@ class UnhandledRequest(Exception):
     Raised by the manager when none of it's providers responds to a request.
     """
     pass
+
+
+class StaleBlockchain(Exception):
+    """
+    Raised by the stalecheck_middleware when the latest block is too old.
+    """
+    def __init__(self, block, allowable_delay):
+        last_block_date = datetime.datetime.fromtimestamp(block.timestamp).strftime('%c')
+        message = (
+            "The latest block, #%d, is %d seconds old, but is only allowed to be %d s old. "
+            "The date of the most recent block is %s. Continue syncing and try again..." %
+            (block.number, time.time() - block.timestamp, allowable_delay, last_block_date)
+        )
+        super().__init__(message, block, allowable_delay)
